# documentation/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Персональная библиотека для атмосферных литературных вечеров. Интерактивная документация: /docs (Swagger) и /redoc.",
    "title": "nocturne API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/stats/insights": {
      "post": {
        "description": "Наблюдения о привычках чтения — по кнопке (тратит токены).\n\nЭтап 9: доступно КАЖДОМУ вошедшему, а не только админу. Статистика личная\n(считается по своей полке), и запрещать её тестерам бессмысленно — иначе\nкнопка у них всегда отвечала бы 403. Расходы держат лимиты частоты\n(rate_limit.py, 20 AI-запросов в час) и капы у провайдеров (з.36).\nНе сохраняем: цифры меняются с каждой прочитанной книгой, и устаревший\nкомментарий хуже, чем его отсутствие.\nСессию открываем вручную КОРОТКИМ отрезком (не через get_session) — дальше\nидёт долгий AI-вызов, держать соединение всё это время не нужно.",
        "operationId": "create_insights_api_v1_stats_insights_post",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Insights",
        "tags": [
          "stats"
        ]
      }
    },
...
  }
}